+++ kubernetes-examples/src/main/java/io/fabric8/kubernetes/examples/BindingExample.java
@@ -1,0 +1,49 @@
+/**
+ * Copyright (C) 2015 Red Hat, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.fabric8.kubernetes.examples;
+
+import io.fabric8.kubernetes.api.model.Binding;
+import io.fabric8.kubernetes.api.model.BindingBuilder;
+import io.fabric8.kubernetes.api.model.KubernetesResourceList;
+import io.fabric8.kubernetes.client.DefaultKubernetesClient;
+import io.fabric8.kubernetes.client.KubernetesClient;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import java.io.IOException;
+
+public class BindingExample {
+  private static Logger logger = LoggerFactory.getLogger(WaitUntilReadyExample.class);
+
+  public static void main(String args[]) throws IOException, InterruptedException {
+    try (final KubernetesClient client = new DefaultKubernetesClient()) {
+      Binding binding = new BindingBuilder()
+        .withNewMetadata().withName("testpod").endMetadata()
+        .withNewTarget().withKind("Node").withApiVersion("v1").withName("minikube").endTarget()
+        .build();
+
+      client.bindings().inNamespace("default").create(binding);
+    }
+  }
+
+  private static void log(String action, Object obj) {
+    logger.info("{}: {}", action, obj);
+  }
+
+  private static void log(String action) {
+    logger.info(action);
+  }
+}